+++ CHANGELOG.md
@@ -10,8 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   to create Memberships and Posts in the test db, while testing django popolo
 
 ### Changed
-- original and normalized Education levels are now storable
-- original and normalized Professions are now storable
+- original and normalized Education levels and Professions names are unique
+- original and normalized Education levels and Professions are now storable
 - Organization has a thematic_classification field
 - add_classification and add_classification_rel methods separated
 - update_links, update_sources, update_contact_details implemented